Better distinguish between edit and play mode tests
