+++ Assets/Tests/EditModeTests/CircularBufferTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 36f5db38dc60cb140ab2486634e9b3f8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
